fix:[BE] change quiz report participants

See merge request s08-final/S08P31A707!346

+++ back/quiz/src/main/java/com/luckquiz/quiz/db/repository/QuizReportCustomRepository.java
@@ -70,7 +70,7 @@ public class QuizReportCustomRepository {
                         quizGuest.id,
                         quizGuest.guestNickname,
                         quizGuest.correctCount.divide(quizGuest.totalCount).longValue(),
-                        quizGuest.score)
+                        quizGuest.score.doubleValue())
                 )
                 .from(quizGuest)
                 .where(
